pyproject.toml
[project]
name = "fakestore-backend-drf"
version = "1.0.0"
description = "A Fakestore REST API built with Django and Django REST Framework."
authors = [
    {name = "Eduardo da Silva", email = "eduardo.silva@gmail.com"},
]
dependencies = [
    "django>=4.2.13",
    "djangorestframework>=3.15.1",
    "django-cors-headers>=4.3.1",
    "django-cloudinary-storage>=0.3.0",
    "passage-identity>=2.5.0",
    "django-extensions>=3.2.3",
    "django-filter>=24.2",
    "drf-spectacular>=0.27.2",
    "setuptools>=70.0.0",
    "python-magic>=0.4.27",
    "Pillow>=10.3.0",
    "dj-database-url>=2.2.0",
    "python-dotenv>=1.0.1",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
format = ["isort>=5.13.2", "black>=24.4.2"]
dev = [
    "black>=24.4.2",
    "isort>=5.13.2",
    "pylint>=3.2.2",
    "netifaces>=0.11.0",
]

[tool.pdm.scripts]
pre_dev = "python ./scripts/set_my_ip.py"
dev = "python manage.py runserver 0.0.0.0:8000"
createsuperuser = "python manage.py createsuperuser"
pre_migrate = "python manage.py makemigrations"
migrate = "python manage.py migrate"
dbshell = "python manage.py dbshell"
shell = "python manage.py shell"
shellp= "python manage.py shell_plus"
test = "python manage.py test"
loaddata = "python manage.py loaddata core.json"
dumpdata = "python manage.py dumpdata --indent 2"

[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
  \.git
 | \.hg
 | \.mypy_cache
 | \.tox
 | \.venv
 | _build
 | buck-out
 | build
 | __pypackages__
 )/
'''
